LayoutTests/imported/w3c/web-platform-tests/custom-state-pseudo-class/state-css-selector-shadow-dom.tentative-expected.txt


PASS state selector has no influence when state is not applied
FAIL state selector has influence when state is applied assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
PASS state selector only applies on given ident
